pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "modular-rag-mcp-server"
version = "0.1.0"
description = "Modular RAG MCP Server — 可扩展、高可观测的智能问答与知识检索框架"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may use real backends)",
    "e2e: End-to-end tests (full pipeline)",
    "slow: Tests that take a long time to run",
]
pythonpath = [".", "src"]
